@@ -2,8 +2,10 @@ package com.baeldung.recordswithjpa;
 
 import com.baeldung.recordswithjpa.entity.Book;
 import com.baeldung.recordswithjpa.repository.BookRepository;
-import org.junit.jupiter.api.AfterEach;
-import org.junit.jupiter.api.BeforeEach;
+
+import org.junit.jupiter.api.AfterAll;
+import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.TestInstance;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 
@@ -11,11 +13,12 @@ import jakarta.transaction.Transactional;
 
 @Transactional
 @SpringBootTest
+@TestInstance(TestInstance.Lifecycle.PER_CLASS)
 public class RecordsAsJpaIntegrationTest {
     @Autowired
     protected BookRepository bookRepository;
 
-    @BeforeEach
+    @BeforeAll
     void setUp() {
 
         Book book = new Book(null,"The Lord of the Rings", "J.R.R. Tolkien", "978-0544003415");
@@ -28,7 +31,7 @@ public class RecordsAsJpaIntegrationTest {
         bookRepository.save(book3);
     }
 
-    @AfterEach
+    @AfterAll
     void tearDown() {
         bookRepository.deleteAll();
     }
